Fixed the initial bugs and added new files
