+++ logging/logging-source/garden_logging.py
@@ -3,23 +3,29 @@ import sys
 import serial
 
 from args_parser import ArgumentParser
+from logger import Logger
 
 def run_logger(argv):
     # Read in command line arguments
     parser = ArgumentParser()
     parser.parse(argv)
     port = parser.port
+    directory = parser.directory
+    file_prefix = parser.file_prefix
+    file_suffix = parser.file_suffix
 
     # Open port
-    s = serial.Serial(port) 
+    
 
     # Begin reading data
-
-
+    #poll the port reader
+    
     # Parse data
+    #pass the data to the data parser
 
 
     # Either create a new file or write to an existing one
+    #pass the results to the logger
  
 
     # On error, attempt to reopen port
@@ -31,7 +37,7 @@ def run_logger(argv):
 
 
 
-    s.close()
+
