+++ logging/logging-source/logger.py
@@ -1,0 +1,4 @@
+
+
+class Logger:
+    def init(self):
